Fix object size initialization
Currently the size of a VMObject, `objectSize`, is set before the
constructor is called on the allocated memory. It is an
error to set a field of an object before the constructor is called.

The C++ standard states in 12.7.1 (N4296):
  For an object with a non-trivial constructor, referring to any
  non-static member or base class of the object before the constructor
  begins execution results in undefined behavior. For an object with a
  non-trivial destructor, referring to any non-static member or base class
  of the object after the destructor finishes execution results in
  undefined behavior.

This results in the C++ compiler eliminating the store to `objectSize`
during the call to `operator new`.  The constructors of VMObjects relies
on `this->objectSize` being appropriately set before the constructor is
called. As a result, `objectSize` must be set in the constructor.

Signed-off-by: Andrew Young <youngar17@gmail.com>